requirements.txt
streamlit
tensorflow
numpy
pandas
matplotlib
seaborn
Pillow
plotly
scikit-learn
streamlit
tensorflow
numpy
pandas
Pillow
matplotlib
plotly
gdown
